refactor: Container ID를 Long에서 Integer로 변경, 불필요한 파라미터 제거
- Container 엔티티의 ID 타입을 Long에서 Integer로 변경
- 모든 관련 레이어(Controller, Service, Repository, DTO, Exception)에서 타입 수정
- 구현되지 않은 컨트롤러 파라미터들 제거
- 통계 및 검색 기능 주석 처리

+++ backend/src/main/java/solid/backend/container/dto/BatchContainerVisibilityDto.java
@@ -18,7 +18,7 @@ public class BatchContainerVisibilityDto {
     
     @NotEmpty(message = "컨테이너 ID 목록은 필수입니다")
     @Size(max = 100, message = "한 번에 최대 100개까지 변경 가능합니다")
-    private List<Long> containerIds; // 변경할 컨테이너 ID 목록
+    private List<Integer> containerIds; // 변경할 컨테이너 ID 목록
     
     @NotNull(message = "공개 상태는 필수입니다")
     private Boolean isPublic; // 변경할 공개 상태 (true: 공개, false: 비공개)
